Move the Icon component to Components/
