+++ src/lib/Components/Bidding/Components/Icon.tsx
