+++ src/lib/Components/Bidding/Screens/RegistrationResult.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import { View } from "react-native"
 
+import { Icon20 } from "../Components/Icon"
 import { Flex } from "../Elements/Flex"
-import { Icon20 } from "../Elements/Icon"
 
 import { BiddingThemeProvider } from "../Components/BiddingThemeProvider"
 import { BidGhostButton } from "../Components/Button"
